+++ Manatee.Json.Tests/DevTest.cs
@@ -11,6 +11,7 @@ using Manatee.Json.Patch;
 using Manatee.Json.Pointer;
 using Manatee.Json.Schema;
 using Manatee.Json.Serialization;
+using Manatee.Json.Tests.Schema;
 using Manatee.Json.Tests.Test_References;
 using NUnit.Framework;
 
@@ -24,8 +25,25 @@ namespace Manatee.Json.Tests
 		[Test]
 		public void Test()
 		{
+			//JsonSchemaOptions.OutputFormat = SchemaValidationOutputFormat.Flag;
+
 			var serializer = new JsonSerializer();
-			Console.WriteLine(serializer.Serialize(MetaSchemas.Draft2019_06).GetIndentedString());
+
+			var schemaFile = @"C:\Users\gregs\Downloads\Sample\fhir.schema.json";
+			var jsonFile = @"C:\Users\gregs\Downloads\Sample\sample.json";
+
+			var schemaText = File.ReadAllText(schemaFile);
+			var schemaJson = JsonValue.Parse(schemaText);
+			var schema = serializer.Deserialize<JsonSchema>(schemaJson);
+
+			var jsonText = File.ReadAllText(jsonFile);
+			var json = JsonValue.Parse(jsonText);
+
+			//var metaResults = schema.ValidateSchema();
+			var results = schema.Validate(json);
+
+			//File.WriteAllText(@"C:\Users\gregs\Downloads\Sample\metaResults.json", serializer.Serialize(metaResults).GetIndentedString());
+			File.WriteAllText(@"C:\Users\gregs\Downloads\Sample\results.json", serializer.Serialize(results).GetIndentedString());
 		}
 	}
 }